Better colour scheme with following prompt:
So I either workout at Gym or Cult. I could have additional bodyweight workouts. In a gym workout, I could either workout back+biceps, chest+triceps, legs or forearms. In a Cult workout, I could either do a HRX, SnC, Burn or Boxing workout in addition to Dance or Yoga. The bodyweight workouts are mostly addons to a cult or gym workout. So I want you to consider this pattern of mine and help create a theme of colours such that on any day, all the workouts done are distinctly visible to a human eye in the colours heatmap.

Signed-off-by: Diptanshu Jamgade <djamgade@nvidia.com>